+++ src/indicators/service.ts
@@ -1,6 +1,11 @@
 import { isMainThread, threadId } from 'worker_threads'
 import utils from '../utils'
 import {
+  createIndicator,
+  feedCandle,
+  getWarmupCandles,
+} from '@gainium/indicators'
+import type {
   RSI,
   MFI,
   ADX,
@@ -34,7 +39,6 @@ import {
   MAR,
   BBPB,
   DIV,
-  DIVUsableOscillators,
   SuperTrend,
   PC,
   ATR,
@@ -54,9 +58,7 @@ import {
   intervalMap,
   StatusEnum,
   IndicatorEnum,
-  MAEnum,
   serviceLogRedis,
-  RangeType,
 } from '../../types'
 import logger from '../utils/logger'
 import type {
@@ -329,468 +331,21 @@ class InternalIndicator {
     exchange: _exchange,
     symbolCode,
   }: IndicatorCreationConfig) {
-    this.length = 0
     this.indicatorName =
       indicatorConfig.type === IndicatorEnum.ma
         ? indicatorConfig.maType
         : indicatorConfig.type
-    if (indicatorConfig.type === IndicatorEnum.st) {
-      this.indicator = new SuperTrend(
-        indicatorConfig.factor,
-        indicatorConfig.atrLength,
-      )
-      this.length = 750
-    }
-    if (indicatorConfig.type === IndicatorEnum.dc) {
-      this.indicator = new DonchianChannels(indicatorConfig.length)
-      this.length = indicatorConfig.length + 1
-    }
-    if (indicatorConfig.type === IndicatorEnum.pp) {
-      this.indicator = new PriorPivot(
-        indicatorConfig.ppHighLeft,
-        indicatorConfig.ppHighRight,
-        indicatorConfig.ppLowLeft,
-        indicatorConfig.ppLowRight,
-        indicatorConfig.ppMult,
-      )
-      this.length =
-        Math.max(
-          indicatorConfig.ppHighLeft + indicatorConfig.ppHighRight,
-          indicatorConfig.ppLowLeft + indicatorConfig.ppLowRight,
-          1000,
-        ) + 1
-    }
-    if (indicatorConfig.type === IndicatorEnum.pc) {
-      this.indicator = new PC(indicatorConfig.pcUp, indicatorConfig.pcDown)
-      this.length = 1
-    }
-    if (indicatorConfig.type === IndicatorEnum.rsi) {
-      this.indicator = new RSI(
-        indicatorConfig.interval,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.atr) {
-      this.indicator = new ATR(indicatorConfig.interval)
-      this.length = indicatorConfig.interval
-    }
-    if (indicatorConfig.type === IndicatorEnum.adr) {
-      this.indicator = new ADR(indicatorConfig.interval)
-      this.length = indicatorConfig.interval
-    }
-    if (indicatorConfig.type === IndicatorEnum.ecd) {
-      this.indicator = new ECD()
-      this.length = 2
-    }
-    if (indicatorConfig.type === IndicatorEnum.ath) {
-      this.indicator = new ATH(indicatorConfig.lookback)
-      this.length = indicatorConfig.lookback + 1
-    }
-    if (indicatorConfig.type === IndicatorEnum.vo) {
-      this.indicator = new VO(
-        indicatorConfig.voShort,
-        indicatorConfig.voLong,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.voLong +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.psar) {
-      this.length = 100
-      this.indicator = new PSAR(
-        indicatorConfig.start,
-        indicatorConfig.inc,
-        indicatorConfig.max,
-      )
-    }
-    if (indicatorConfig.type === IndicatorEnum.div) {
-      this.length = 200
-      this.indicator = new DIV(
-        indicatorConfig.oscillators.map((v) =>
-          v.toLowerCase(),
-        ) as DIVUsableOscillators[],
-        indicatorConfig.leftBars ?? 3,
-        indicatorConfig.rightBars ?? 1,
-        indicatorConfig.rangeLower ?? 1,
-        indicatorConfig.rangeUpper ?? 60,
-      )
-    }
-    if (indicatorConfig.type === IndicatorEnum.mom) {
-      this.length = Math.max(
-        indicatorConfig.interval +
-          (indicatorConfig.percentile
-            ? +(indicatorConfig.percentileLookback ?? '0')
-            : 0),
-        100,
-      )
-      this.indicator = new MOM(
-        indicatorConfig.interval,
-        //@ts-ignore
-        indicatorConfig.source,
-        indicatorConfig.percentile,
-        indicatorConfig.percentileLookback,
-        indicatorConfig.percentilePercentage,
-      )
-    }
-    if (indicatorConfig.type === IndicatorEnum.mar) {
-      this.indicator = new MAR(
-        indicatorConfig.mar1type,
-        indicatorConfig.mar1length,
-        indicatorConfig.mar2type,
-        indicatorConfig.mar2length,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-        indicatorConfig.trendFilter,
-        indicatorConfig.trendFilterLookback,
-        indicatorConfig.trendFilterValue,
-        indicatorConfig.trendFilterType,
-      )
-      this.length =
-        Math.max(indicatorConfig.mar1length, indicatorConfig.mar2length) +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.mfi) {
-      this.indicator = new MFI(
-        indicatorConfig.interval,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.adx) {
-      this.indicator = new ADX(
-        indicatorConfig.interval,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.bbw) {
-      const bb = new BollingerBands(
-        indicatorConfig.interval,
-        indicatorConfig.bbwMult ?? 2,
-        indicatorConfig.bbwMa ?? MAEnum.sma,
-        indicatorConfig.bbwMaLength ?? 20,
-      )
-      this.indicator = new BollingerBandsWidth(
-        bb,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.bbwMaLength ?? 0) *
-          (indicatorConfig.bbwMa === MAEnum.tema
-            ? 3
-            : indicatorConfig.bbwMa === MAEnum.dema
-              ? 2
-              : 1) +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.bbpb) {
-      const bb = new BollingerBands(
-        indicatorConfig.interval,
-        indicatorConfig.bbwMult ?? 2,
-        indicatorConfig.bbwMa ?? MAEnum.sma,
-        indicatorConfig.bbwMaLength ?? 20,
-      )
-      this.indicator = new BBPB(
-        bb,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.bbwMaLength ?? 0) *
-          (indicatorConfig.bbwMa === MAEnum.tema
-            ? 3
-            : indicatorConfig.bbwMa === MAEnum.dema
-              ? 2
-              : 1) +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.kcpb) {
-      const kc = new KeltnerChannel(
-        indicatorConfig.interval,
-        indicatorConfig.multiplier ?? 2,
-        indicatorConfig.ma ?? MAEnum.ema,
-        indicatorConfig.range ?? RangeType.atr,
-        indicatorConfig.rangeLength ?? 10,
-      )
-      this.indicator = new KeltnerChannelPB(
-        kc,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.rangeLength ?? 10) +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.kc) {
-      this.indicator = new KeltnerChannel(
-        indicatorConfig.interval,
-        indicatorConfig.multiplier ?? 2,
-        indicatorConfig.ma ?? MAEnum.ema,
-        indicatorConfig.range ?? RangeType.atr,
-        indicatorConfig.rangeLength ?? 10,
-      )
-      this.length =
-        indicatorConfig.interval + (indicatorConfig.rangeLength ?? 10) + 1
-    }
-    if (indicatorConfig.type === IndicatorEnum.bbwp) {
-      const bb = new BollingerBands(indicatorConfig.interval, 1, MAEnum.sma, 20)
-      this.indicator = new BBWP(bb, indicatorConfig.lookback)
-      this.length = indicatorConfig.interval + indicatorConfig.lookback
-    }
-    if (indicatorConfig.type === IndicatorEnum.bb) {
-      this.indicator = new BollingerBands(
-        indicatorConfig.interval,
-        indicatorConfig.bbwMult ?? 2,
-        indicatorConfig.bbwMa ?? MAEnum.sma,
-        indicatorConfig.bbwMaLength ?? 20,
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.bbwMaLength ?? 0) *
-          (indicatorConfig.bbwMa === MAEnum.tema
-            ? 3
-            : indicatorConfig.bbwMa === MAEnum.dema
-              ? 2
-              : 1)
-    }
-    if (indicatorConfig.type === IndicatorEnum.macd) {
-      const maSource = indicatorConfig.maSource === MAEnum.sma ? SMA : EMA
-      const maSignal = indicatorConfig.maSignal === MAEnum.sma ? SMA : EMA
-      this.indicator = new MACD(
-        new maSource(indicatorConfig.shortInterval),
-        new maSource(indicatorConfig.longInterval),
-        new maSignal(indicatorConfig.signalInterval),
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        Math.max(indicatorConfig.longInterval + indicatorConfig.shortInterval) +
-        indicatorConfig.signalInterval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.stoch) {
-      this.indicator = new StochasticOscillator(
-        indicatorConfig.k,
-        indicatorConfig.ksmooth,
-        indicatorConfig.dsmooth,
-      )
-      this.length =
-        indicatorConfig.k + indicatorConfig.ksmooth + indicatorConfig.dsmooth
-    }
-    if (indicatorConfig.type === IndicatorEnum.cci) {
-      this.indicator = new CCI(
-        indicatorConfig.interval,
-        'hlc3',
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.ao) {
-      this.indicator = new AO(
-        5,
-        34,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        34 +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.stochRSI) {
-      this.indicator = new StochasticRSI(
-        indicatorConfig.interval,
-        indicatorConfig.k,
-        indicatorConfig.ksmooth,
-        indicatorConfig.dsmooth,
-      )
-      this.length =
-        indicatorConfig.interval +
-        indicatorConfig.k +
-        indicatorConfig.ksmooth +
-        indicatorConfig.dsmooth
-    }
-    if (indicatorConfig.type === IndicatorEnum.wr) {
-      this.indicator = new WilliamsR(
-        indicatorConfig.interval,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.bullBear) {
-      this.indicator = new BullBear(indicatorConfig.interval)
-      this.length = indicatorConfig.interval
-    }
-    if (indicatorConfig.type === IndicatorEnum.uo) {
-      this.indicator = new UltimateOscillator(
-        indicatorConfig.fast,
-        indicatorConfig.middle,
-        indicatorConfig.slow,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        Math.max(
-          indicatorConfig.fast,
-          indicatorConfig.middle,
-          indicatorConfig.slow,
-        ) +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.ic) {
-      this.indicator = new IchimokuCloud(
-        indicatorConfig.conversionPeriods,
-        indicatorConfig.basePeriods,
-        indicatorConfig.laggingSpan2Periods,
-        indicatorConfig.laggingSpan,
-      )
-      this.length = Math.max(
-        indicatorConfig.conversionPeriods,
-        indicatorConfig.basePeriods,
-        indicatorConfig.laggingSpan2Periods,
-        indicatorConfig.laggingSpan,
-      )
-    }
-    if (indicatorConfig.type === IndicatorEnum.ma) {
-      if (indicatorConfig.maType === MAEnum.ema) {
-        this.indicator = new EMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.sma) {
-        this.indicator = new SMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.wma) {
-        this.indicator = new WMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.hma) {
-        this.indicator = new HMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.vwma) {
-        this.indicator = new VWMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.dema) {
-        this.indicator = new DEMA(indicatorConfig.interval)
-        this.length = 2 * indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.tema) {
-        this.indicator = new TEMA(indicatorConfig.interval)
-        this.length = 3 * indicatorConfig.interval + 300
-      }
-      if (indicatorConfig.maType === MAEnum.rma) {
-        this.indicator = new RMA(indicatorConfig.interval)
-        this.length = indicatorConfig.interval + 300
-      }
-    }
-    if (indicatorConfig.type === IndicatorEnum.tv) {
-      this.indicator = new TVTA(
-        indicatorConfig.checkLevel,
-        indicatorConfig.useAsEntryExitPoints,
-      )
-      this.length = 3000
-    }
-    if (indicatorConfig.type === IndicatorEnum.sr) {
-      this.indicator = new SupportResistance(
-        indicatorConfig.leftBars,
-        indicatorConfig.rightBars,
-      )
-      this.length = indicatorConfig.leftBars + indicatorConfig.rightBars
-    }
-    if (indicatorConfig.type === IndicatorEnum.mfi) {
-      this.indicator = new MFI(
-        indicatorConfig.interval,
-        indicatorConfig.percentile,
-        +(indicatorConfig.percentileLookback ?? '0'),
-        +(indicatorConfig.percentilePercentage ?? '0'),
-      )
-      this.length =
-        indicatorConfig.interval +
-        (indicatorConfig.percentile
-          ? +(indicatorConfig.percentileLookback ?? '0')
-          : 0)
-    }
-    if (indicatorConfig.type === IndicatorEnum.qfl) {
-      this.indicator = new QFL(
-        indicatorConfig.basePeriods,
-        indicatorConfig.pumpPeriods,
-        indicatorConfig.pump,
-        indicatorConfig.baseCrack,
-      )
-      this.length = indicatorConfig.basePeriods + indicatorConfig.pumpPeriods
-    }
-    if (indicatorConfig.type === IndicatorEnum.obfvg) {
-      this.indicator = new OBFVG()
-      this.length = 1000
-    }
-    if (indicatorConfig.type === IndicatorEnum.lw) {
-      this.indicator = new LongWick(
-        indicatorConfig.lwThreshold ?? 2,
-        indicatorConfig.lwMaxDuration ?? 1000,
-      )
-      this.length = 201
+    // Indicator instantiation + warmup are now sourced from
+    // @gainium/indicators (createIndicator / getWarmupCandles) so the live
+    // service, the archive's on-demand evaluator, and any future consumer
+    // share one definition of "what config produces what instance, with
+    // what warmup."
+    const created = createIndicator(indicatorConfig)
+    if (created) {
+      this.indicator = created
+      this.length = getWarmupCandles(indicatorConfig)
+    } else {
+      this.length = 0
     }
 
     this.type = indicatorConfig.type
@@ -1028,9 +583,22 @@ class InternalIndicator {
       old.c = +c
       old.v = +v
     }
-    const startIndex = start - (!forceOld ? this.period : 0)
+    const exchangesWithClosedOnly = [
+      ExchangeEnum.binance,
+      ExchangeEnum.bybit,
+      ExchangeEnum.okx,
+    ]
+    const isClosedOnly = exchangesWithClosedOnly.some(
+      (e) =>
+        this.exchange.toUpperCase().includes(e.toUpperCase()) ||
+        e.toUpperCase().includes(this.exchange.toUpperCase()),
+    )
+    // closed-only streams deliver `start` as the just-closed candle's own start
+    // (legacy streams use the next-period open), so they index like `forceOld`.
+    const closedCandleAlreadyKnown = forceOld || isClosedOnly
+    const startIndex = closedCandleAlreadyKnown ? start : start - this.period
     if (
-      ((start > this.start && this.start !== 0) || forceOld) &&
+      ((start > this.start && this.start !== 0) || isClosedOnly || forceOld) &&
       !this.updateCandlesHistory.has(startIndex)
     ) {
       this.updateCandlesHistory.add(startIndex)
@@ -1090,7 +658,7 @@ class InternalIndicator {
           true,
         )
       }
-      this.start = +start + (forceOld ? this.period : 0)
+      this.start = +start + (closedCandleAlreadyKnown ? this.period : 0)
       this.lastCandle = {
         open: `${old.o}`,
         high: `${old.h}`,
@@ -1121,92 +689,15 @@ class InternalIndicator {
     ) {
       return
     }
-    if (this.indicator && this.indicator instanceof VO) {
-      this.indicator.next(+value.v)
-    }
-    if (
-      this.indicator &&
-      (this.indicator instanceof RSI ||
-        this.indicator instanceof MACD ||
-        this.indicator instanceof EMA ||
-        this.indicator instanceof DEMA ||
-        this.indicator instanceof TEMA ||
-        this.indicator instanceof RMA ||
-        this.indicator instanceof SMA ||
-        this.indicator instanceof WMA ||
-        this.indicator instanceof HMA)
-    ) {
-      this.indicator?.next(+value.c)
-    }
-    if (
-      this.indicator &&
-      (this.indicator instanceof ADX ||
-        this.indicator instanceof StochasticOscillator ||
-        this.indicator instanceof CCI ||
-        this.indicator instanceof StochasticRSI ||
-        this.indicator instanceof WilliamsR ||
-        this.indicator instanceof UltimateOscillator ||
-        this.indicator instanceof IchimokuCloud ||
-        this.indicator instanceof SupportResistance ||
-        this.indicator instanceof QFL ||
-        this.indicator instanceof PSAR ||
-        this.indicator instanceof SuperTrend ||
-        this.indicator instanceof ATR ||
-        this.indicator instanceof PriorPivot ||
-        this.indicator instanceof ADR ||
-        this.indicator instanceof ATH)
-    ) {
-      this.indicator?.next({
-        high: +value.h,
-        low: +value.l,
-        close: +value.c,
-      })
-    }
-    if (
-      this.indicator &&
-      (this.indicator instanceof BullBear ||
-        this.indicator instanceof MOM ||
-        this.indicator instanceof ECD ||
-        this.indicator instanceof DonchianChannels ||
-        this.indicator instanceof OBFVG ||
-        this.indicator instanceof LongWick)
-    ) {
-      this.indicator?.next({
-        high: +value.h,
-        low: +value.l,
-        close: +value.c,
-        open: +value.o,
-      })
-    }
-    if (
-      this.indicator &&
-      (this.indicator instanceof VWMA ||
-        this.indicator instanceof TVTA ||
-        this.indicator instanceof MFI ||
-        this.indicator instanceof BollingerBandsWidth ||
-        this.indicator instanceof BollingerBands ||
-        this.indicator instanceof KeltnerChannel ||
-        this.indicator instanceof KeltnerChannelPB ||
-        this.indicator instanceof BBWP ||
-        this.indicator instanceof BBPB ||
-        this.indicator instanceof MAR ||
-        this.indicator instanceof DIV ||
-        this.indicator instanceof PC)
-    ) {
-      this.indicator?.next({
-        high: +value.h,
-        low: +value.l,
-        close: +value.c,
-        open: +value.o,
-        volume: +value.v,
-      })
-    }
-    if (this.indicator && this.indicator instanceof AO) {
-      this.indicator?.next({
-        high: +value.h,
-        low: +value.l,
-      })
-    }
+    // Per-class instanceof dispatch lives in @gainium/indicators' feedCandle —
+    // one source of truth for "which input shape goes to which indicator."
+    feedCandle(this.indicator, {
+      o: +value.o,
+      h: +value.h,
+      l: +value.l,
+      c: +value.c,
+      v: +value.v,
+    })
     try {
       const result = this.indicator ? this.indicator.result : undefined
       if (result !== undefined && result !== null) {